refactor: move _util to utils
