+++ weasel/cli/clone.py
@@ -6,8 +6,7 @@ from typing import Optional
 from wasabi import msg
 
 from .. import about
-from .._util import get_git_version, git_checkout, git_repo_branch_exists
-from ..utils import ensure_path
+from ..utils import ensure_path, get_git_version, git_checkout, git_repo_branch_exists
 from .main import COMMAND, PROJECT_FILE, Arg, Opt, app
 
 DEFAULT_REPO = about.__projects__
