+++ weasel/tests/test_cli.py
@@ -3,13 +3,13 @@ import time
 
 import pytest
 import srsly
+from confection import ConfigValidationError
 
-from weasel._util import ConfigValidationError, is_subpath_of, load_project_config
-from weasel._util import substitute_project_variables, validate_project_commands
 from weasel.cli.remote_storage import RemoteStorage
 from weasel.cli.run import _check_requirements
 from weasel.schemas import ProjectConfigSchema, validate
-from weasel.utils import make_tempdir
+from weasel.utils import is_subpath_of, load_project_config, make_tempdir
+from weasel.utils import substitute_project_variables, validate_project_commands
 
 
 def test_issue11235():
