+++ weasel/utils/__init__.py
@@ -1,6 +1,11 @@
-from .commands import join_command, run_command, split_command
+from .commands import download_file, get_checksum, get_git_version, get_hash
+from .commands import git_checkout, git_repo_branch_exists, join_command, run_command
+from .commands import split_command, upload_file, validate_project_commands
+from .config import load_project_config, parse_config_overrides
+from .config import substitute_project_variables
 from .environment import ENV_VARS, check_bool_env_var, check_spacy_env_vars
-from .filesystem import ensure_path, is_cwd, make_tempdir, working_dir
+from .filesystem import ensure_path, ensure_pathy, is_cwd, is_subpath_of, make_tempdir
+from .filesystem import working_dir
 from .logging import logger
 from .modules import import_file
 from .structures import SimpleFrozenDict, SimpleFrozenList
