+++ weasel/utils/commands/validation.py
@@ -1,0 +1,83 @@
+import sys
+from configparser import InterpolationError
+from contextlib import contextmanager
+from pathlib import Path
+from typing import TYPE_CHECKING, Any, Dict, Optional, Union
+
+from confection import ConfigValidationError
+from wasabi import msg
+
+from weasel.cli.main import COMMAND, PROJECT_FILE
+
+if TYPE_CHECKING:
+    pass
+
+
+@contextmanager
+def show_validation_error(
+    file_path: Optional[Union[str, Path]] = None,
+    *,
+    title: Optional[str] = None,
+    desc: str = "",
+    show_config: Optional[bool] = None,
+    hint_fill: bool = True,
+):
+    """Helper to show custom config validation errors on the CLI.
+
+    file_path (str / Path): Optional file path of config file, used in hints.
+    title (str): Override title of custom formatted error.
+    desc (str): Override description of custom formatted error.
+    show_config (bool): Whether to output the config the error refers to.
+    hint_fill (bool): Show hint about filling config.
+    """
+    try:
+        yield
+    except ConfigValidationError as e:
+        title = title if title is not None else e.title
+        if e.desc:
+            desc = f"{e.desc}" if not desc else f"{e.desc}\n\n{desc}"
+        # Re-generate a new error object with overrides
+        err = e.from_error(e, title="", desc=desc, show_config=show_config)
+        msg.fail(title)
+        print(err.text.strip())
+        if hint_fill and "value_error.missing" in err.error_types:
+            config_path = (
+                file_path
+                if file_path is not None and str(file_path) != "-"
+                else "config.cfg"
+            )
+            msg.text(
+                "If your config contains missing values, you can run the 'init "
+                "fill-config' command to fill in all the defaults, if possible:",
+                spaced=True,
+            )
+            print(f"{COMMAND} init fill-config {config_path} {config_path} \n")
+        sys.exit(1)
+    except InterpolationError as e:
+        msg.fail("Config validation error", e, exits=1)
+
+
+def validate_project_commands(config: Dict[str, Any]) -> None:
+    """Check that project commands and workflows are valid, don't contain
+    duplicates, don't clash  and only refer to commands that exist.
+
+    config (Dict[str, Any]): The loaded config.
+    """
+    command_names = [cmd["name"] for cmd in config.get("commands", [])]
+    workflows = config.get("workflows", {})
+    duplicates = set([cmd for cmd in command_names if command_names.count(cmd) > 1])
+    if duplicates:
+        err = f"Duplicate commands defined in {PROJECT_FILE}: {', '.join(duplicates)}"
+        msg.fail(err, exits=1)
+    for workflow_name, workflow_steps in workflows.items():
+        if workflow_name in command_names:
+            err = f"Can't use workflow name '{workflow_name}': name already exists as a command"
+            msg.fail(err, exits=1)
+        for step in workflow_steps:
+            if step not in command_names:
+                msg.fail(
+                    f"Unknown command specified in workflow '{workflow_name}': {step}",
+                    f"Workflows can only refer to commands defined in the 'commands' "
+                    f"section of the {PROJECT_FILE}.",
+                    exits=1,
+                )
